Added test GetCalculationGridAreasAsync_UseProcessManagerFeature
